Move validation to route parameters
