+++ src/DependencyInjection/CQRSExtension.php
@@ -13,7 +13,7 @@ use DigitalCraftsman\CQRS\RequestDataTransformer\RequestDataTransformerInterface
 use DigitalCraftsman\CQRS\RequestDecoder\RequestDecoderInterface;
 use DigitalCraftsman\CQRS\RequestValidator\RequestValidatorInterface;
 use DigitalCraftsman\CQRS\ResponseConstructor\ResponseConstructorInterface;
-use DigitalCraftsman\CQRS\Routing\RoutePayload;
+use DigitalCraftsman\CQRS\Routing\RouteParameters;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -88,45 +88,45 @@ final class CQRSExtension extends Extension
          */
         $config = $this->processConfiguration($configuration, $configs);
 
-        RoutePayload::validateRequestValidatorClasses(
+        RouteParameters::validateRequestValidatorClasses(
             $config['query_controller']['default_request_validator_classes'],
             null,
         );
-        RoutePayload::validateRequestDecoderClass($config['query_controller']['default_request_decoder_class']);
-        RoutePayload::validateRequestDataTransformerClasses(
+        RouteParameters::validateRequestDecoderClass($config['query_controller']['default_request_decoder_class']);
+        RouteParameters::validateRequestDataTransformerClasses(
             $config['query_controller']['default_request_data_transformer_classes'],
             null,
         );
-        RoutePayload::validateDTOConstructorClass($config['query_controller']['default_dto_constructor_class']);
-        RoutePayload::validateDTOValidatorClasses(
+        RouteParameters::validateDTOConstructorClass($config['query_controller']['default_dto_constructor_class']);
+        RouteParameters::validateDTOValidatorClasses(
             $config['query_controller']['default_dto_validator_classes'],
             null,
         );
-        RoutePayload::validateHandlerWrapperClasses(
+        RouteParameters::validateHandlerWrapperClasses(
             $config['query_controller']['default_handler_wrapper_classes'],
             null,
         );
-        RoutePayload::validateResponseConstructorClass($config['query_controller']['default_response_constructor_class']);
+        RouteParameters::validateResponseConstructorClass($config['query_controller']['default_response_constructor_class']);
 
-        RoutePayload::validateRequestValidatorClasses(
+        RouteParameters::validateRequestValidatorClasses(
             $config['command_controller']['default_request_validator_classes'],
             null,
         );
-        RoutePayload::validateRequestDecoderClass($config['command_controller']['default_request_decoder_class']);
-        RoutePayload::validateRequestDataTransformerClasses(
+        RouteParameters::validateRequestDecoderClass($config['command_controller']['default_request_decoder_class']);
+        RouteParameters::validateRequestDataTransformerClasses(
             $config['command_controller']['default_request_data_transformer_classes'],
             null,
         );
-        RoutePayload::validateDTOConstructorClass($config['command_controller']['default_dto_constructor_class']);
-        RoutePayload::validateDTOValidatorClasses(
+        RouteParameters::validateDTOConstructorClass($config['command_controller']['default_dto_constructor_class']);
+        RouteParameters::validateDTOValidatorClasses(
             $config['command_controller']['default_dto_validator_classes'],
             null,
         );
-        RoutePayload::validateHandlerWrapperClasses(
+        RouteParameters::validateHandlerWrapperClasses(
             $config['command_controller']['default_handler_wrapper_classes'],
             null,
         );
-        RoutePayload::validateResponseConstructorClass($config['command_controller']['default_response_constructor_class']);
+        RouteParameters::validateResponseConstructorClass($config['command_controller']['default_response_constructor_class']);
 
         foreach ($config['query_controller'] as $key => $value) {
             $container->setParameter('cqrs.query_controller.'.$key, $value);
